Puts each of the panels in the simulation tab into its own file
