+++ NuRadioReco/eventbrowser/default_layout.py
@@ -4,3 +4,18 @@ default_layout= {
     },
     'showlegend': True
 }
+
+efield_plot_colors = [
+    ['rgb(91, 179, 240)', 'rgb(31, 119, 180)'],
+    ['rgb(255, 187, 94)', 'rgb(255, 127, 14)'],
+    ['rgb(104, 220, 104)', 'rgb(44, 160, 44)'],
+    ['rgb(255, 99, 100)', 'rgb(214, 39, 40)'],
+    ['rgb(208, 163, 249)', 'rgb(148, 103, 189)'],
+    ['rgb(200, 146, 135)', 'rgb(140, 86, 75)']
+]
+efield_plot_linestyles = {
+    'direct': 'solid',
+    'reflected': 'dash',
+    'refracted': 'dot'
+}
+polarizaiton_names = ['r', 'theta', 'phi']
